Fix cannot "Run" Qt project built with cmake
This commit amends 651ea0. Add QTDIR to PATH when starting from a
debug configuration but without debbuging enabled (Ctrl+F5) or from
a release configuration.

When a native session is starting, check if QTDIR is defined as
environment variable and, if so, add a reference to it to PATH.

Task-number: QTVSADDINBUG-1264
Change-Id: I8d2879210e2de9882b1db09933e173118679bfdb
Reviewed-by: Miguel Costa <miguel.costa@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>